Add assets: main_background and previews
