+++ src/App.js
@@ -2,6 +2,7 @@ import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import Auth from './pages/Auth/Auth';
 import Home from './pages/Home';
+import Main from './pages/Main/Main';
 
 function App() {
   return (
@@ -9,7 +10,7 @@ function App() {
       <Routes>
         <Route path='/' element={<Home />} />
         <Route path='/auth' element={<Auth />} />
-        <Route path='/main/:id' element={<Auth />} />
+        <Route path='/main' element={<Main />} />
       </Routes>
     </div>
   );
