+++ src/pages/Main/main.css
@@ -1,0 +1,3 @@
+.main {
+  background: url('../../assets/backgrounds/background_main.png');
+}
